authentication_project/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.0
psycopg2==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
uv==0.2.5
